requirements.txt
flask
flask-cors
python-decouple
python-dotenv
flask_marshmallow
requests
pandas